+++ .github/workflows/ci.yml
@@ -27,7 +27,7 @@ jobs:
       - uses: ./.github/actions/cache  # avoids sporadic 500s from deno’s CDN
       - uses: denoland/setup-deno@v1   # using ourself to install deno could compromise the tests
       - run: deno task test --coverage=cov_profile
-      - run: deno coverage cov_profile --lcov --output=cov_profile.lcov
+      - run: deno coverage cov_profile --lcov --exclude=tests/ --output=cov_profile.lcov
       - uses: coverallsapp/github-action@v1
         with:
           path-to-lcov: cov_profile.lcov
