+++ deno.jsonc
@@ -4,7 +4,8 @@
     "strict": true
   },
   "tasks": {
-    "test": "deno test --allow-net --allow-read --allow-env --allow-run --allow-write --unstable",
+    "test": "deno test --allow-net --allow-read --allow-env --allow-run --allow-write --unstable --import-map=import-map.test.json",
+    "coverage" : "scripts/run_coverage.sh",
     "typecheck": "deno check --unstable ./src/app.ts",
     // runs this source checkout for testing
     //NOTE doesn’t work :/ passes the args to deno and doesn’t run the script
