+++ src/hooks/index.ts
@@ -1,22 +1,24 @@
 // order is important to avoid circular dependencies and thus uncaught ReferenceErrors
 
-import usePrefix from "./usePrefix.ts"
-import useOffLicense from "./useOffLicense.ts"
-import useDownload from "./useDownload.ts"
-import useCache from "./useCache.ts"
-import useCellar from "./useCellar.ts"
-import useExec from "./useExec.ts"
-import useFetch from "./useFetch.ts"
-import useFlags from "./useFlags.ts"
-import useInventory from "./useInventory.ts"
-import useShellEnv from "./useShellEnv.ts"
-import usePantry from "./usePantry.ts"
-import useVirtualEnv from "./useVirtualEnv.ts"
-import usePackageYAML, { usePackageYAMLFrontMatter } from "./usePackageYAML.ts"
-import useSync from "./useSync.ts"
-import useVersion from "./useVersion.ts"
-import useMoustaches from "./useMoustaches.ts"
-import useErrorHandler from "./useErrorHandler.ts"
+import usePrefix from "hooks/usePrefix.ts"
+import useOffLicense from "hooks/useOffLicense.ts"
+import useDownload from "hooks/useDownload.ts"
+import useCache from "hooks/useCache.ts"
+import useCellar from "hooks/useCellar.ts"
+import useExec from "hooks/useExec.ts"
+import useFetch from "hooks/useFetch.ts"
+import useFlags from "hooks/useFlags.ts"
+import useInventory from "hooks/useInventory.ts"
+import useShellEnv from "hooks/useShellEnv.ts"
+import usePantry from "hooks/usePantry.ts"
+import useVirtualEnv from "hooks/useVirtualEnv.ts"
+import usePackageYAML, { usePackageYAMLFrontMatter } from "hooks/usePackageYAML.ts"
+import useSync from "hooks/useSync.ts"
+import useVersion from "hooks/useVersion.ts"
+import useMoustaches from "hooks/useMoustaches.ts"
+import useErrorHandler from "hooks/useErrorHandler.ts"
+import usePrint from "hooks/usePrint.ts"
+import useRun from "hooks/useRun.ts"
 
 // but we can sort these alphabetically
 export {
@@ -38,4 +40,6 @@ export {
   useSync,
   useVersion,
   useVirtualEnv,
+  usePrint,
+  useRun,
 }
