+++ src/hooks/usePrint.ts
@@ -1,0 +1,9 @@
+const encoder = new TextEncoder()
+
+export default function usePrint() {
+  const print = (msg: string) => Deno.stdout.write(encoder.encode(`${msg}\n`))
+
+  return {
+    print 
+  }
+}
